cache seems to be working, removal is working, loading is working, need to clean up the interfaces and get rid of the cache being reliant on the seat service being passed in

+++ src/main/java/model/SeatHold.java
@@ -4,17 +4,17 @@ import java.util.ArrayList;
 
 public class SeatHold {
 
-    private long seatHoldId;
+    private Integer seatHoldId;
     private String customerEmail;
     private ArrayList<Seat> heldSeats;
 
-    public SeatHold(long seatHoldId, String customerEmail, ArrayList<Seat> heldSeats){
+    public SeatHold(Integer seatHoldId, String customerEmail, ArrayList<Seat> heldSeats){
         this.seatHoldId = seatHoldId;
         this.customerEmail = customerEmail;
         this.heldSeats = heldSeats;
     }
 
-    public long getSeatHoldId(){
+    public Integer getSeatHoldId(){
         return this.seatHoldId;
     }
